saling goods view, order lists, order card, theme,

+++ src/main/java/com/second_hand_trading_platform/second_hand_trading_platform/pojo/entity/goods/Goods.java
@@ -9,9 +9,9 @@ public class Goods {
     Integer id;
     String name;
     String info;
-    int price;
+    Double price;
     String ownerId;
-    int originalPrice;
+    Double originalPrice;
     String defaultImage;
     int count;
     int newPercentage;
